% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_api.R
\name{geocode_response}
\alias{geocode_response}
\title{Function to get coordinates given an address, from geocode API
Authors: Maria Treesa Sebastian(marse306), Brian Masinde(brima748), Omkar (omkbh878)}
\arguments{
\item{address}{as vector}
}
\value{
latitude and longitude of the address as a matirx
}
\description{
Function to get coordinates given an address, from geocode API
Authors: Maria Treesa Sebastian(marse306), Brian Masinde(brima748), Omkar (omkbh878)
}
