% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installing_packages.R
\name{InstallPackages}
\alias{InstallPackages}
\title{This R script is used for installing  and loading the required packages}
\value{
nothing
}
\description{
This R script is used for installing  and loading the required packages
}
