% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_access_api.R
\name{reverse_geocode_response}
\alias{reverse_geocode_response}
\title{Function to return address given latitude and longitude}
\usage{
reverse_geocode_response(lat, long, map_key)
}
\arguments{
\item{lat}{as numeric}

\item{long}{as numeric}

\item{map_key}{as google api key}
}
\value{
a formatted address
}
\description{
Function to return address given latitude and longitude
}
\author{
: Maria,Masinde,Omkar
}
